프로젝트 : Invitation 화면 및 기능 중간 커밋
