+++ spring-app/src/main/java/kr/co/co_working/member/controller/MemberApiController.java
@@ -43,7 +43,7 @@ public class MemberApiController {
      * @return
      * @throws Exception
      */
-    @GetMapping("/api/v1/MemberList")
+    @GetMapping("/api/v1/memberList")
     public ResponseDto<?> readMemberList(@RequestBody MemberRequestDto.READ dto) throws Exception {
         return ResponseDto.ofSuccess("멤버목록 조회에 성공했습니다.", service.readMemberList(dto));
     }
@@ -54,7 +54,7 @@ public class MemberApiController {
      * @return
      * @throws Exception
      */
-    @GetMapping("/api/v1/workspace/MemberList-in-workspace")
+    @GetMapping("/api/v1/workspace/memberList-in-workspace")
     public ResponseDto<?> readMemberListInWorkspace(@ModelAttribute WorkspaceRequestDto.READ dto) throws Exception {
         return ResponseDto.ofSuccess("멤버목록 조회에 성공했습니다.", service.readMemberListInWorkspace(dto));
     }
@@ -65,7 +65,7 @@ public class MemberApiController {
      * @return
      * @throws Exception
      */
-    @GetMapping("/api/v1/workspace/MemberList-not-in-workspace")
+    @GetMapping("/api/v1/workspace/memberList-not-in-workspace")
     public ResponseDto<?> readMemberListNotInWorkspace(@ModelAttribute WorkspaceRequestDto.READ dto) throws Exception {
         return ResponseDto.ofSuccess("멤버목록 조회에 성공했습니다.", service.readMemberListNotInWorkspace(dto));
     }
